Refactor: 게시물 카테고리, 콘텐츠 블럭 컴포넌트 분리 및 파일 구조 수정
