+++ src/app/admin/notice/create/_components/CategoryDropdown.css.ts
@@ -1,0 +1,10 @@
+import { BodyBold } from '@/styles/font.css';
+import { style } from '@vanilla-extract/css';
+
+export const dropdown = style([
+  BodyBold,
+  {
+    padding: '0.5rem',
+    borderRadius: '8px',
+  },
+]);
