+++ src/app/admin/notice/create/_components/block/index.tsx
@@ -1,0 +1,8 @@
+import BodyContent from './BodyContent';
+import SubTitleContent from './SubTitleContet';
+import ButtonContent from './ButtonContent';
+import ImageContent from './ImageContent';
+import LineContent from './LineContent';
+import NoteContent from './NoteContent';
+
+export { BodyContent, SubTitleContent, ButtonContent, ImageContent, LineContent, NoteContent };
